more backend work

+++ ClientSide/taskmanager/src/actions/projectAction.js
@@ -42,9 +42,11 @@ export const getProject = (id, history) => async dispatch => {
 };
 
 export const deleteProject = id => async dispatch => {
-    await axios.delete(`http://localhost:8080/api/project/${id}`);
-    dispatch({
-        type: DELETE_PROJECT,
-        payload: id
-    })
+    if(window.confirm("Are you sure you want to delete this task?")){
+        await axios.delete(`http://localhost:8080/api/project/${id}`);
+        dispatch({
+            type: DELETE_PROJECT,
+            payload: id
+        })
+    }
 }
